nesgo/instructions: Add instruction type
The instruction struct represents a given instruction. It contains all
relevant metadata, as well as the function that knows how to execute
that particular function.

The CPU will read an instruction's opcode form memory, use the
instructions slice to fetch the corresponding instruction, and call
it's Run function to execute it, passing in all the relevant state.

Signed-off-by: Duarte Nunes <duarte.m.nunes@gmail.com>