nesgo/instructions: Add the position type
The position struct encapsulates the addresses resolved from a
particular addressing mode. In particular, it returns a value from
memory, if the addressing mode resolved to an address in memory, or
the value in the accumulator, if the addressing mode was
"accumulator". Conversely, it knows how to write a value to the
correct location.

Signed-off-by: Duarte Nunes <duarte.m.nunes@gmail.com>